argoslabs/microsoft/o365/requirements.txt
# pip dependent packages

# default for alabs
alabs.common

# ...
# for Office 365

O365>=2.0.5
